src/main/java/Item.java
public class Item implements ISell {


    private String name;
    private double buyingPrice;
    private double sellingPrice;

    public Item(String name, double buyingPrice, double sellingPrice) {
        this.name = name;
        this.sellingPrice = sellingPrice;
        this.buyingPrice = buyingPrice;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public double getBuyingPrice() {
        return buyingPrice;
    }

    public void setBuyingPrice(double buyingPrice) {
        this.buyingPrice = buyingPrice;
    }

    public double getSellingPrice() {
        return sellingPrice;
    }

    public void setSellingPrice(double sellingPrice) {
        this.sellingPrice = sellingPrice;
    }

    public double calculateMarkup(){
        double result = this.sellingPrice -= this.buyingPrice;
        return -result;
    }
}
